+++ src/MockIO/T/Process.hs
@@ -2,30 +2,22 @@ module MockIO.T.Process
   ( tests )
 where
 
+import Base1T  hiding  ( toList )
+
 -- base --------------------------------
 
-import Data.Eq        ( Eq )
 import Data.Foldable  ( any )
-import Data.Function  ( ($), (&) )
+import Data.Function  ( flip )
 import Data.Maybe     ( isJust )
-import Data.Word      ( Word8 )
 import GHC.Exts       ( IsList( toList ) )
-import System.Exit    ( ExitCode )
-import System.IO      ( IO )
-import Text.Show      ( Show )
-
--- base-unicode-symbols ----------------
-
-import Data.Function.Unicode  ( (∘) )
-import Data.Monoid.Unicode    ( (⊕) )
 
 -- containers-plus ---------------------
 
 import ContainersPlus.Insert  ( (⨭) )
 
--- data-textual ------------------------
+-- env-plus ----------------------------
 
-import Data.Textual  ( Printable, toString, toText )
+import Env.Types  ( ӭ, ә )
 
 -- exceptions --------------------------
 
@@ -38,7 +30,6 @@ import FPath.AbsFile  ( AbsFile, absfile )
 -- lens --------------------------------
 
 import Control.Lens.Fold  ( (^?) )
-import Control.Lens.Lens  ( Lens' )
 
 -- log-plus ----------------------------
 
@@ -61,54 +52,36 @@ import MockIO.MockIOClass  ( MockIOClass )
 
 -- monaderror-io -----------------------
 
-import MonadError           ( ѥ )
-import MonadError.IO.Error  ( AsIOError( _IOError ), ioeErrorString
+import MonadError.IO.Error  ( IOError, ioeErrorString
                             , ioeFilename, ioeLocation, isNoSuchThingError )
 
 -- monadio-plus ------------------------
 
-import MonadIO                        ( MonadIO )
 import MonadIO.Error.CreateProcError  ( ProcError )
 import MonadIO.Error.ProcExitError    ( ProcExitError, _ProcExitError
                                       , stdErr, stdOut )
 import MonadIO.File                   ( devnull )
 import MonadIO.Process.CmdSpec        ( CmdArgs( CmdArgs ), CmdExe( CmdExe )
                                       , cmdArgs, cmdExe, expExitVal, mkCmd )
+import MonadIO.Process.ExitInfo       ( ExitInfo )
 import MonadIO.Process.ExitStatus     ( ExitStatus( ExitVal ), exitVal )
-
--- more-unicode ------------------------
-
-import Data.MoreUnicode.Bool     ( pattern 𝕿 )
-import Data.MoreUnicode.Either   ( 𝔼 )
-import Data.MoreUnicode.Functor  ( (⊳), (⊳⊳) )
-import Data.MoreUnicode.Lens     ( (⊣) )
-import Data.MoreUnicode.Maybe    ( pattern 𝕵 )
-import Data.MoreUnicode.Monad    ( (≫) )
-import Data.MoreUnicode.Natural  ( ℕ )
-import Data.MoreUnicode.String   ( 𝕊 )
-import Data.MoreUnicode.Text     ( 𝕋 )
 
 -- mtl ---------------------------------
 
-import Control.Monad.Except  ( MonadError )
+import Control.Monad.Reader  ( runReaderT )
 
 -- prettyprinter -----------------------
 
 import Prettyprinter              ( layoutCompact )
 import Prettyprinter.Render.Text  ( renderStrict )
 
--- tasty -------------------------------
-
-import Test.Tasty  ( TestName, TestTree, testGroup )
-
 -- tasty-hunit -------------------------
 
-import Test.Tasty.HUnit  ( Assertion, (@=?), assertBool, assertEqual, testCase )
+import Test.Tasty.HUnit  ( Assertion, assertBool, assertEqual )
 
 -- tasty-plus --------------------------
 
-import TastyPlus  ( (≟), assertIOError, assertJust, assertRight, runTestsP
-                  , runTestsReplay, runTestTree )
+import TastyPlus  ( (≟), assertIOError, assertJust )
 
 -- text --------------------------------
 
@@ -119,17 +92,13 @@ import Data.Text  ( isInfixOf, unlines, unpack )
 import qualified  Data.Text.ICU
 import Data.Text.ICU  ( Regex, find, regex )
 
--- tfmt --------------------------------
-
-import Text.Fmt  ( fmt )
-
 ------------------------------------------------------------
 --                     local imports                      --
 ------------------------------------------------------------
 
 import qualified  MockIOPlus.Paths  as  Paths
 
-import MockIO.Process                ( system )
+import MockIO.Process                ( ꙩ, system )
 import MockIO.Process.CmdRW          ( CmdRW( CmdR ) )
 import MockIO.Process.MLMakeIStream  ( MLMakeIStream )
 
@@ -143,30 +112,30 @@ foo = unlines [ "jimmy 7"
 
 grep_ ∷ (MonadIO μ, MLMakeIStream σ,
          MonadError ProcError μ, MonadLog (Log MockIOClass) μ) ⇒
-        [𝕋] → σ → μ (ExitStatus, (𝕋,𝕋))
+        [𝕋] → σ → μ (ExitInfo, (𝕋,𝕋))
 grep_ args input =
   let cmd = mkCmd Paths.grep args & expExitVal ⨭ 1
    in system Notice CmdR (ExitVal 1,("","")) input cmd NoMock
 
 grep ∷ (MonadIO μ, MLMakeIStream σ, MonadError ProcError μ) ⇒
-       𝕋 → σ → μ ((ExitStatus, (𝕋,𝕋)), Log MockIOClass)
+       𝕋 → σ → μ ((ExitInfo, (𝕋,𝕋)), Log MockIOClass)
 grep pat input = runPureLoggingT $ grep_ [pat] input
 
 grepaf ∷ (MonadIO μ, MonadError ProcError μ) ⇒
-       𝕋 → AbsFile → μ ((ExitStatus, (𝕋,𝕋)), Log MockIOClass)
+       𝕋 → AbsFile → μ ((ExitInfo, (𝕋,𝕋)), Log MockIOClass)
 grepaf pat fn = runPureLoggingT $ devnull ≫ grep_ [pat, toText fn]
 
 -- for repl use
 
 {- | grep a pattern from some `Text`; capture the logs -}
 _grep_ ∷ (MonadIO μ, MLMakeIStream σ) ⇒
-        𝕋 → σ → μ (𝔼 ProcError ((ExitStatus, (𝕋,𝕋)), Log MockIOClass))
+        𝕋 → σ → μ (𝔼 ProcError ((ExitInfo, (𝕋,𝕋)), Log MockIOClass))
 _grep_ pat input = ѥ $ grep pat input
 
 {- | grep a pattern from some `Text`; write the logs to stderr -}
 
 _grep ∷ (MonadIO μ, MonadMask μ, MLMakeIStream σ) ⇒
-        𝕋 → σ → μ (𝔼 ProcError (ExitStatus, (𝕋,𝕋)))
+        𝕋 → σ → μ (𝔼 ProcError (ExitInfo, (𝕋,𝕋)))
 _grep pat input = logit $ grep_ [pat] input
 
 {- | Perform a list of tests independently against the result of an IO.
@@ -179,13 +148,13 @@ ioTests nm s xs =
 
 ----------------------------------------
 
-data ProcResult = ProcResult { exit ∷ ExitStatus
+data ProcResult = ProcResult { exit ∷ ExitInfo
                              , out  ∷ 𝕋
                              , err  ∷ 𝕋
                              , log  ∷ Log MockIOClass
                              }
 
-mkProcResult ∷ ((ExitStatus, (𝕋,𝕋)), Log MockIOClass) → ProcResult
+mkProcResult ∷ ((ExitInfo, (𝕋,𝕋)), Log MockIOClass) → ProcResult
 mkProcResult ((ex,(ot,er)),lg) = ProcResult ex ot er lg
 
 {- | Test the results of an external process.  Note that the proc is run
@@ -199,7 +168,7 @@ testProc ∷ TestName
          → [(TestName,[LogEntry MockIOClass] → Assertion)]
          → TestTree
 testProc nm s expExit expOut expErr expLog =
-  ioTests nm s $ [ ("exit",   (\ r → ExitVal expExit ≟ exit r))
+  ioTests nm s $ [ ("exit",   (\ r → ExitVal expExit ≟ exit r ⊣ exitVal))
                  , ("stdout", (\ r → expOut  ≟ out r))
                  , ("stderr", (\ r → expErr  ≟ err r))
                  ]
@@ -234,11 +203,13 @@ tests =
     p t = mkProcResult ⊳⊳ (ѥ @ProcError $ grep t foo)
     logTests ∷ Word8 → [(TestName,[LogEntry MockIOClass] → Assertion)]
     logTests e = [ ([fmt|exit %d|] e, logIncludes ([fmt|Execution exit %d|] e))
-               , ("cmd grep", logIncludes $ [fmt|<CMD> %T|] Paths.grep)
-               , ("provided text",
-                  logMatches $
-                    regex [] "using provided Text «jimmy.*» as input stream")
-               ]
+                 , ("cmd grep",
+                    logMatches $
+                      regex [] $ "<CMD> «Pid \\d+» " ⊕ toText Paths.grep)
+                 , ("provided text",
+                    logMatches $
+                      regex [] "using provided Text «jimmy.*» as input stream")
+                 ]
   in
     testGroup
       "Process"
@@ -322,5 +293,12 @@ _tests = runTestsP tests
 
 _testr ∷ 𝕊 → ℕ → IO ExitCode
 _testr = runTestsReplay tests
+
+
+-- development helpers ---------------------------------------------------------
+
+-- _xx = devnull ≫ (¡ ([absfile|/grep|],[]))
+_xx ∷ IO (𝔼 IOError (𝔼 ProcError (ExitInfo, ())))
+_xx = logit $  splitMError $ flip runReaderT NoMock $ ꙩ ([absfile|/run/current-system/sw/bin/grep|],["martyn","/etc/group"]::[𝕋],[ӭ (ә "HOME")])
 
 -- that's all, folks! ----------------------------------------------------------